Import npreal2 "for Linux 5.x" on 20191018.
Actually, the difference between the codebases "for 3.x", "4.x" and
"5.x" is inessential; the codebase "for 2.x" differs greatly, but
it is for legacy kernel versions and is therefore not considered here.
`mxloadsvr.c' appears to differ quite a little between "3.x" and "4.x",
but the resulting behaviours are still the same.

SHA256 checksums for (where $i is 3, 4 or 5, respectively)
`moxa-real-tty-drivers-for-linux-$i.x.x-driver-v$i.0.tgz':
5ea1dc83825d14eb0ce2e59726a27e6a2da271397a3ac198a9c55628d41928ac
b8a75609ccd1457649debf8f26a1c20d9e7e0bd6419276e01e772045d75cce28
33da5d4b1ff9853e9d58c7905f1fdf09a3e284658f42437210155c4c913f4dad